💄 Refactor modal styles
