+++ components/Modal/ClaimRewardModal/RewardItem.module.scss
@@ -1,45 +1,3 @@
-.claimRewardModal {
-  max-width: 300px;
-  width: 100%;
-  padding: 30px;
-  color: $black;
-  background-color: $white;
-  border-radius: $radius-sm;
-
-  @include responsive(SM) {
-    width: 300px;
-  }
-}
-
-.header {
-  @include flexbox(between);
-  margin-bottom: 20px;
-}
-
-.closeButton {
-  @include flexbox;
-  width: 40px;
-  height: 40px;
-  margin-right: -8px;
-  color: $gray-4;
-  transition: $transition-short;
-
-  &:hover {
-    color: $gray-3;
-  }
-
-  svg {
-    width: 24px;
-    height: 24px;
-  }
-}
-
-.title {
-  @include textStyle(18, 700);
-
-  color: $black;
-}
-
 .detail {
   margin-bottom: 20px;
 }
