--- components/Modal/ConnectModal.module.scss
@@ -1,34 +1,0 @@
-.connectModal {
-  max-width: 300px;
-  padding: 30px;
-  background-color: $white;
-  border-radius: $radius-sm;
-}
-
-.title {
-  @include textStyle(18, 700);
-  margin-bottom: 20px;
-  color: $black;
-}
-
-.agreement {
-  @include flexbox(between, start);
-  @include textStyle(14, 400);
-  margin-bottom: 20px;
-  color: $black;
-
-  .checkbox {
-    top: 1px;
-    flex-shrink: 0;
-    margin-right: 10px;
-  }
-
-  a {
-    margin: 0 0.2em;
-    text-decoration: underline;
-
-    &:hover {
-      opacity: 0.5;
-    }
-  }
-}
