--- components/Modal/MetaMaskGuideModal.module.scss
@@ -1,40 +1,0 @@
-.metamaskGuideModal {
-  max-width: 320px;
-  padding: 30px;
-  background-color: $white;
-  border-radius: $radius-sm;
-}
-
-.header {
-  @include flexbox(between);
-  margin-bottom: 20px;
-}
-
-.closeButton {
-  @include flexbox;
-  width: 40px;
-  height: 40px;
-  margin-right: -8px;
-  color: $gray-4;
-  transition: $transition-short;
-
-  &:hover {
-    color: $gray-3;
-  }
-
-  svg {
-    width: 24px;
-    height: 24px;
-  }
-}
-
-.title {
-  @include textStyle(18, 700);
-  color: $black;
-}
-
-.desc {
-  @include textStyle(14, 400);
-  margin-bottom: 20px;
-  color: $black;
-}
