+++ components/Modal/PoolModals/Composition.module.scss
@@ -1,0 +1,58 @@
+.details {
+  @include flexbox(start, start);
+  flex-direction: column;
+  width: 100%;
+  overflow: hidden;
+  background-color: $gray-1;
+  border-radius: $radius-sm;
+
+  &.reverse {
+    flex-direction: column-reverse;
+  }
+}
+
+.detailItem {
+  @include flexbox(between);
+  flex-wrap: wrap;
+  width: 100%;
+  padding: 8px;
+  background-color: $gray-1;
+  border-radius: $radius-sm;
+
+  dt {
+    @include flexbox(start);
+
+    strong {
+      @include textStyle(14, 700);
+      text-transform: uppercase;
+    }
+  }
+
+  dd {
+    @include flexbox(center, end);
+    @include textStyle(16, 700);
+    flex-direction: column;
+    text-align: right;
+  }
+
+  .usd {
+    @include inlineFlexbox(start);
+    @include textStyle(12, 500);
+    margin-top: 0px;
+    padding-right: 4px;
+    color: $gray-3;
+  }
+
+  .percent {
+    @include textStyle(12, 500);
+    margin-left: 4px;
+    color: $gray-4;
+  }
+}
+
+.token {
+  width: 28px;
+  height: 28px;
+  margin-right: 8px;
+  color: $white;
+}
