+++ components/Modal/PoolModals/ExitPreviewModal.tsx
@@ -1,7 +1,7 @@
 import { MouseEvent, useCallback, useEffect, useState } from 'react'
 import { AnimatePresence } from 'framer-motion'
 import type { Event } from 'ethers'
-import styles from './style.module.scss'
+import styles from '../style.module.scss'
 
 import { ModalCategory } from 'app/states/modal'
 import { TxAction } from 'services/transaction'
