+++ components/Modal/PoolModals/Warning.module.scss
@@ -1,0 +1,29 @@
+.warning {
+  position: relative;
+  padding: 10px 8px 10px 40px;
+  margin: 16px 0;
+  background-color: lighten($red, 42);
+  border: 1px solid lighten($red, 35);
+  border-radius: $radius-sm;
+
+  h3 {
+    @include textStyle(14, 700);
+    flex-direction: column;
+    margin-bottom: 4px;
+    line-height: 1.3;
+
+    svg {
+      position: absolute;
+      top: 10px;
+      left: 10px;
+      width: 20px;
+      height: 20px;
+      color: $red;
+    }
+  }
+
+  p {
+    @include textStyle(13);
+    line-height: 1.3;
+  }
+}
