--- components/Modal/WithdrawPreviewModal.module.scss
@@ -1,183 +1,0 @@
-.withdrawPreviewModal {
-  max-width: 320px;
-  width: calc(100% - 20px);
-  padding: 30px;
-  color: $black;
-  background-color: $white;
-  border-radius: $radius-sm;
-}
-
-.header {
-  @include flexbox(between);
-  margin-bottom: 20px;
-}
-
-.closeButton {
-  @include flexbox;
-  width: 40px;
-  height: 40px;
-  margin-right: -8px;
-  color: $gray-4;
-  transition: $transition-short;
-
-  &:hover {
-    color: $gray-3;
-  }
-
-  svg {
-    width: 24px;
-    height: 24px;
-  }
-}
-
-.title {
-  @include textStyle(18, 700);
-  color: $black;
-}
-
-.subtitle {
-  @include textStyle(14, 700);
-  display: block;
-  margin-bottom: 10px;
-}
-
-.withdrawDetail {
-  @include flexbox(between);
-  @include textStyle(16, 700);
-  width: 100%;
-  margin-bottom: 20px;
-
-  dt {
-    @include flexbox(start);
-
-    .token {
-      &.wncg {
-        margin-left: -4px;
-      }
-    }
-  }
-
-  dd {
-    @include flexbox(center, end);
-    flex-direction: column;
-    text-align: right;
-  }
-
-  .usd {
-    justify-content: flex-end;
-  }
-}
-
-.tokens {
-  @include inlineFlexbox(start);
-  color: $white;
-}
-
-.token {
-  @include inlineFlexbox;
-  flex-grow: 0;
-  flex-shrink: 0;
-  width: 22px !important;
-  height: 22px !important;
-  border: 1px solid $white;
-}
-
-.usd {
-  @include inlineFlexbox(start);
-  @include textStyle(12, 500);
-  flex-grow: 1;
-  color: $gray-3;
-
-  svg {
-    width: 0.8em;
-    height: 0.8em;
-    margin-right: 0.2em;
-    margin-bottom: -0.005em;
-  }
-}
-
-.claimDetail {
-  margin-bottom: 20px;
-}
-
-.detailItem {
-  @include flexbox(between);
-  flex-wrap: wrap;
-  width: 100%;
-  padding: 4px 0;
-  margin-bottom: 4px;
-  color: $black;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-
-  dt {
-    @include flexbox(start);
-    @include textStyle(16, 700);
-
-    .token {
-      margin-right: 4px;
-    }
-
-    strong {
-      @include textStyle(14, 700);
-    }
-  }
-
-  dd {
-    @include flexbox(center, end);
-    flex-direction: column;
-    text-align: right;
-
-    .token {
-      width: 24px;
-      height: 24px;
-    }
-
-    &.isBig {
-      width: 100%;
-      margin-top: 8px;
-    }
-  }
-}
-
-.toast {
-  padding: 12px 16px 14px;
-  background-color: $white;
-
-  .header {
-    @include flexbox(start);
-    margin-bottom: 2px;
-
-    .title {
-      @include textStyle(14, 500);
-      color: $black;
-    }
-  }
-
-  .desc {
-    @include textStyle(13, 400);
-    color: $gray-5;
-  }
-
-  .closeButton {
-    @include flexbox;
-    position: absolute;
-    top: 12px;
-    right: 10px;
-    width: 20px;
-    height: 20px;
-    color: darken($gray-3, 5);
-    transition: $transition-short;
-
-    &:hover {
-      color: $gray-5;
-    }
-
-    svg {
-      width: 18px;
-      height: 18px;
-    }
-  }
-}
