+++ components/Modal/WithdrawPreviewModal/RewardsSummary.module.scss
@@ -1,0 +1,45 @@
+.claimDetail {
+  margin-bottom: 20px;
+}
+
+.detailItem {
+  @include flexbox(between);
+  flex-wrap: wrap;
+  width: 100%;
+  padding: 4px 0;
+  margin-bottom: 4px;
+  color: $black;
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+
+  dt {
+    @include flexbox(start);
+    @include textStyle(16, 700);
+
+    .token {
+      margin-right: 4px;
+    }
+
+    strong {
+      @include textStyle(14, 700);
+    }
+  }
+
+  dd {
+    @include flexbox(center, end);
+    flex-direction: column;
+    text-align: right;
+
+    .token {
+      width: 24px;
+      height: 24px;
+    }
+
+    &.isBig {
+      width: 100%;
+      margin-top: 8px;
+    }
+  }
+}
