+++ components/Modal/WithdrawPreviewModal/WithdrawSummary.module.scss
@@ -1,0 +1,54 @@
+.withdrawDetail {
+  @include flexbox(between);
+  @include textStyle(16, 700);
+  width: 100%;
+  margin-bottom: 20px;
+
+  dt {
+    @include flexbox(start);
+
+    .token {
+      &.wncg {
+        margin-left: -4px;
+      }
+    }
+  }
+
+  dd {
+    @include flexbox(center, end);
+    flex-direction: column;
+    text-align: right;
+  }
+
+  .usd {
+    justify-content: flex-end;
+  }
+}
+
+.tokens {
+  @include inlineFlexbox(start);
+  color: $white;
+}
+
+.token {
+  @include inlineFlexbox;
+  flex-grow: 0;
+  flex-shrink: 0;
+  width: 22px !important;
+  height: 22px !important;
+  border: 1px solid $white;
+}
+
+.usd {
+  @include inlineFlexbox(start);
+  @include textStyle(12, 500);
+  flex-grow: 1;
+  color: $gray-3;
+
+  svg {
+    width: 0.8em;
+    height: 0.8em;
+    margin-right: 0.2em;
+    margin-bottom: -0.005em;
+  }
+}
