+++ styles/constants/_radius.scss
@@ -1,2 +1,3 @@
 $radius-sm: 4px;
 $radius-md: 8px;
+$radius-lg: 16px;
