+++ styles/constants/_typography.scss
@@ -21,6 +21,10 @@ $font-size-18: 18px;
 $line-height-18: 24px;
 $letter-spacing-18: -0.03em;
 
+$font-size-20: 20px;
+$line-height-20: 28px;
+$letter-spacing-20: -0.03em;
+
 $font-size-24: 24px;
 $line-height-24: 32px;
 $letter-spacing-24: -0.03em;
