+++ styles/mixins/_text-style.scss
@@ -16,6 +16,9 @@
   } @else if ($scale == 18) {
     font-size: typo.$font-size-18;
     line-height: typo.$line-height-18;
+  } @else if ($scale == 20) {
+    font-size: typo.$font-size-20;
+    line-height: typo.$line-height-20;
   } @else if ($scale == 24) {
     font-size: typo.$font-size-24;
     line-height: typo.$line-height-24;
@@ -41,6 +44,8 @@
       letter-spacing: typo.$letter-spacing-16 - 0.01em;
     } @else if ($scale == 18) {
       letter-spacing: typo.$letter-spacing-18 - 0.01em;
+    } @else if ($scale == 20) {
+      letter-spacing: typo.$letter-spacing-20 - 0.01em;
     } @else if ($scale == 24) {
       letter-spacing: typo.$letter-spacing-24 - 0.01em;
     } @else if ($scale == 30) {
@@ -57,6 +62,8 @@
       letter-spacing: typo.$letter-spacing-16;
     } @else if ($scale == 18) {
       letter-spacing: typo.$letter-spacing-18;
+    } @else if ($scale == 20) {
+      letter-spacing: typo.$letter-spacing-20;
     } @else if ($scale == 24) {
       letter-spacing: typo.$letter-spacing-24;
     } @else if ($scale == 30) {
